[hexalog] track libklio changes
when we receive values for sensors we do not know yet, we cache the
values and ask the sensor for it's name. if we don't receive a name, the
readings are dropped.

we also now cache sensors to avoid reading the entire klio timezone
database every time a reading is received.