implement key rollover
key rollover can happen for many different reasons.

if the current frame counter exceeds a rollover limit, go to the next
key and make sure that enough keys are reserved and valid for future
use.

if all devices in the network agree that a list of keys is obsolete,
remove those, ensure that enough reserved keys exists and possibly move
to the first key in the list if the current outgoing key has been
superseded.

if some device uses a key that is not the first in the list, jump to
that key and ensure that enough reserved keys exist. devices should then
follow and also use only the highest-numbered key, so that the next
rollovers will remove all keys below the now current key.